# Optimize ORC StructSelectiveStreamReader.getRetainedSizeInBytes